+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
  * removed `--all` option from `puli config`
  * changed `puli config` to display all user and effective values by default
  * removed `--force` option from `puli build`
+ * added `--force` option to `puli bind`
 
 * 1.0.0-beta2 (2015-01-27)